fix: renamed function names & store edited data as json
