+++ src/controllers/users/getDashboardSettings.ts
@@ -1,9 +1,9 @@
 import { Request, Response, NextFunction } from 'express';
-import { getDashboardVars } from '../../services';
+import { getDashboardSettings } from '../../services';
 
 export default async (_req: Request, res: Response, next: NextFunction): Promise<void> => {
   try {
-    const data = await getDashboardVars();
+    const data = await getDashboardSettings();
 
     res.json({ data });
   } catch (error) {
