+++ src/routes/users/index.ts
@@ -13,8 +13,8 @@ import {
   matchUserContent,
   changePassword,
   getUserStatistics,
-  editDashboardVars,
-  getDashboardVars,
+  editDashboardSettings,
+  getDashboardSettings,
 } from '../../controllers';
 import {
   constants,
@@ -24,7 +24,7 @@ import {
   createUserSchema,
   matchUserContentSchema,
   changePasswordSchema,
-  editDashboardVarsSchema,
+  editSystemSettingsSchema,
 } from '../../helpers';
 import { checkUserRole } from '../../middleware';
 
@@ -50,8 +50,8 @@ router.patch('/approve/:userId', validator.params(idSchema), approveUser);
 router.patch('/match-user-content', validator.body(matchUserContentSchema), matchUserContent);
 
 router.use(checkUserRole([MASTER_ADMIN]));
-router.get('/dashboard-vars', getDashboardVars);
+router.get('/dashboard-settings', getDashboardSettings);
+router.patch('/edit-dashboard-settings', validator.body(editSystemSettingsSchema), editDashboardSettings);
 router.patch('/block-user/:userId', validator.params(idSchema), blockUser);
-router.patch('/edit-dashboard-vars', validator.body(editDashboardVarsSchema), editDashboardVars);
 
 export default router;
