+++ src/services/user/index.ts
@@ -4,5 +4,5 @@ export { default as getContents } from './getContents';
 export { default as getUsers } from './getUsers';
 export { default as matchUserContent } from './matchUserContent';
 export { default as getNumberOfContent } from './getContentReport';
-export { default as editDashboardVars } from './editDashboardVars';
-export { default as getDashboardVars } from './getDashboardVars';
+export { default as editDashboardSettings } from './editDashboardSettings';
+export { default as getDashboardSettings } from './getDashboardSettings';
